Move ImplQueryInterface to Common.Lib
